work in progress..
* the methodology used in this commit does NOT work
  - I fully expected that it would
  - it does NOT
  - this commit is being tagged so I can point to it later,
    if asked: "why didn't you just...?"
* the methodology is to:
  - detect requests that solve catchas made to issuing domains
    * as was done previously
  - rather than make the issuing request then and there,
    wait until the corresponding call to "handleHeadersReceived"
* problem:
  - there (apparently) is no corresponding call to "handleHeadersReceived"
* to do:
  - add new hook for call to "chrome.webRequest.onCompleted"
  - add plumbing for all providers to receive this new hook function
  - ...hope it receives a corresponding call

- - - -

hCaptcha should send issuing requests from a different hook function

similar to the way redemption works (but in reverse),
detect solved captcha on issuing domains before requests are sent,
but delay the processing of hits until after the response is received

previously, a timeout was used to delay processing;
but this methodology was inexact and could lead to failed attempts.

note that this is in stark contrast to the Cloudflare provider,
which cancels the requests that it detects,
and processes those hits immediately without any delay.