hCaptcha sends issuing requests from 'onCompleted' hook function
similar to the way redemption works (but in reverse),
detect solved captcha on issuing domains before requests are sent,
but delay the processing of hits until after the response is received

previously, a timeout was used to delay processing;
but this methodology was inexact and could lead to failed attempts.

note that this is in stark contrast to the Cloudflare provider,
which cancels the requests that it detects,
and processes those hits immediately without any delay.